+++ packages/vue/index.d.ts
@@ -1,5 +1,17 @@
 import type { ParserOptions } from '@vue/compiler-dom'
+import type { App } from 'vue'
 import type { FilterPattern, Plugin } from 'vite'
+
+// TODO: Make `vue` a peer dep for types in next breaking change
+declare module 'whyframe:app' {
+  interface CreateAppOptions {
+    /**
+     * **[Vue]** Hook to update the app, e.g. adding global components, before
+     * mounting the app to the element.
+     */
+    enhanceApp: (app: App) => void
+  }
+}
 
 export interface Options {
   include?: FilterPattern