+++ packages/vue/src/shared.js
@@ -95,8 +95,11 @@ ${notTemplateCode}`
 import { createApp as _createApp } from 'vue'
 import App from '${entryComponentId}'
 
-export function createApp(el) {
+export function createApp(el, opts) {
   const app = _createApp(App)
+  if (opts?.enhanceApp) {
+    opts.enhanceApp(app)
+  }
   app.mount(el)
   return {
     destroy: () => app.unmount()